emoji-kitchen-scan/pyproject.toml
[tool.poetry]
name = "cheese-coffee-db-gen"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
emojis = "^0.6.0"
tqdm = "^4.62.3"
requests = "^2.27.1"
progressbar2 = "^4.0.0"
alive-progress = "^2.2.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
